share tablecontents going forward
